Update consumables Menu

+++ database/migrations/2023_05_24_125605_create_configs_table.php
@@ -25,7 +25,7 @@ class CreateConfigsTable extends Migration
             $table->string('uom');//set,box,package
             $table->string('consist_of')->nullable();
             
-            $table->integer('price_include_ppn');
+            $table->unsignedBigInteger('price_include_ppn');
             $table->timestamps();
 
            // $table->foreign('unit_id')->references('id')->on('unit');
